+++ apps/api/src/app.module.ts
@@ -1,16 +1,18 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
-import { validateEnvironment } from './config/env.validation';
+import { appConfig, databaseConfig, validateEnv } from './config';
+import { PlatformModule } from './platform';
 
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
-      validate: validateEnvironment,
+      validate: validateEnv,
+      load: [appConfig, databaseConfig],
     }),
+    PlatformModule,
   ],
   controllers: [AppController],
-  providers: [],
 })
 export class AppModule {}